moves bodyarea_lifts tests to their own file
